Deals with the unfair/incorrect points calculation when posts are deleted
When a post that run for sometime with signature is deleted, the old behavior was that this post is still counted as one of the user’s “posts with sig” and so it still earns post points. This is not fair because people can keep posting and deleting posts to increase their points.

Solution:
- Keep track of post deletions
- When counting posts with sig, subtract the number of deleted posts from the count